// clients/js/src/generated/accounts/creditReservation.ts
export type CreditReservation = {
  discriminator: ReadonlyUint8Array;
  version: number;
  coachClientCredits: Address;
  coachAuthority: Address;
  clientWallet: Address;
  bookingId: Array<number>;
  scheduledStartAt: bigint;
  earlyReturnUntil: bigint;
  status: CreditReservationStatus;
  reservedAt: bigint;
  resolvedAt: Option<bigint>;
  bump: number;
  reserved: Array<number>;
};